requirements.txt
openai>=1.102.0
python-dotenv>=1.1.1
requests>=2.31.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python-headless>=4.12.0.88
numpy>=1.26.0
# pdf2image>=1.17.0   # <- only if you need PDFs → images
